+++ lti_consumer/plugin/views.py
@@ -880,6 +880,7 @@ class LtiNrpsContextMembershipViewSet(viewsets.ReadOnlyModelViewSet):
 
         try:
             data = compat.get_course_members(course_key)
+            compat.merge_course_forum_roles(course_key, data)
             self.attach_external_user_ids(data)
 
             # build correct format for the serializer